+++ src/main/java/com/ddevus/currencyExchange/entity/ExchangeRate.java
@@ -1,19 +1,21 @@
 package com.ddevus.currencyExchange.entity;
 
+import java.math.BigDecimal;
+
 public class ExchangeRate {
 
     private int id;
     private Currency baseCurrency;
     private Currency targetCurrency;
-    private float rate;
+    private BigDecimal rate;
 
-    public ExchangeRate(Currency baseCurrency, Currency targetCurrency, float rate) {
+    public ExchangeRate(Currency baseCurrency, Currency targetCurrency, BigDecimal rate) {
         this.baseCurrency = baseCurrency;
         this.targetCurrency = targetCurrency;
         this.rate = rate;
     }
 
-    public ExchangeRate(int id, Currency baseCurrency, Currency targetCurrency, float rate) {
+    public ExchangeRate(int id, Currency baseCurrency, Currency targetCurrency, BigDecimal rate) {
         this.id = id;
         this.baseCurrency = baseCurrency;
         this.targetCurrency = targetCurrency;
@@ -44,11 +46,11 @@ public class ExchangeRate {
         this.targetCurrency = targetCurrency;
     }
 
-    public float getRate() {
+    public BigDecimal getRate() {
         return rate;
     }
 
-    public void setRate(float rate) {
+    public void setRate(BigDecimal rate) {
         this.rate = rate;
     }
